requirements_v1.txt

langchain
langchain-core
pinecone-client
cohere
transformers
python-dotenv
fastapi
uvicorn[standard]
requests
streamlit
